+++ lab/beacon-root-assumption/safe/src/UnusedEthereumAddress.sol
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.20;
+
+contract UnusedEthereumAddress {
+    address public constant ETHEREUM_BEACON_ROOTS = 0x000F3df6D732807Ef1319fB7B8bB8522d0Beac02;
+}